Split out Working Day Helper service
